Move more methods over
