+++ Content.Server/GameObjects/Components/Body/BodyManagerComponent.Parts.cs
@@ -1,21 +1,55 @@
 ﻿#nullable enable
+using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Content.Server.Body;
 using Content.Server.GameObjects.EntitySystems;
 using Content.Server.Interfaces.GameObjects.Components.Interaction;
+using Content.Shared.Body.Part.Properties.Movement;
+using Content.Shared.Body.Part.Properties.Other;
 using Content.Shared.GameObjects.Components.Body;
 using Content.Shared.GameObjects.Components.Damage;
+using Content.Shared.GameObjects.Components.Movement;
 using Robust.Shared.GameObjects.Systems;
 using Robust.Shared.Interfaces.GameObjects;
 using Robust.Shared.Log;
 using Robust.Shared.Utility;
+using Robust.Shared.ViewVariables;
 
 namespace Content.Server.GameObjects.Components.Body
 {
     public partial class BodyManagerComponent
     {
+        private readonly Dictionary<string, IBodyPart> _parts = new Dictionary<string, IBodyPart>();
+
+        /// <summary>
+        ///     All <see cref="IBodyPart"></see> with <see cref="LegProperty"></see>
+        ///     that are currently affecting move speed, mapped to how big that leg
+        ///     they're on is.
+        /// </summary>
+        [ViewVariables]
+        private readonly Dictionary<IBodyPart, float> _activeLegs = new Dictionary<IBodyPart, float>();
+
+        /// <summary>
+        ///     Maps <see cref="BodyTemplate"/> slot name to the <see cref="IBodyPart"/>
+        ///     object filling it (if there is one).
+        /// </summary>
+        [ViewVariables]
+        public IReadOnlyDictionary<string, IBodyPart> Parts => _parts;
+
+        /// <summary>
+        ///     List of all occupied slots in this body, taken from the values of
+        ///     <see cref="Parts"/>.
+        /// </summary>
+        public IEnumerable<string> OccupiedSlots => Parts.Keys;
+
+        /// <summary>
+        ///     List of all slots in this body, taken from the keys of
+        ///     <see cref="Template"/> slots.
+        /// </summary>
+        public IEnumerable<string> AllSlots => Template.Slots.Keys;
+
         public bool TryAddPart(string slot, DroppedBodyPartComponent part, bool force = false)
         {
             DebugTools.AssertNotNull(part);
@@ -368,5 +402,140 @@ namespace Content.Server.GameObjects.Components.Body
             return toReturn;
         }
 
+        private void CalculateSpeed()
+        {
+            if (!Owner.TryGetComponent(out MovementSpeedModifierComponent? playerMover))
+            {
+                return;
+            }
+
+            float speedSum = 0;
+            foreach (var part in _activeLegs.Keys)
+            {
+                if (!part.HasProperty<LegProperty>())
+                {
+                    _activeLegs.Remove(part);
+                }
+            }
+
+            foreach (var (key, value) in _activeLegs)
+            {
+                if (key.TryGetProperty(out LegProperty? leg))
+                {
+                    // Speed of a leg = base speed * (1+log1024(leg length))
+                    speedSum += leg.Speed * (1 + (float) Math.Log(value, 1024.0));
+                }
+            }
+
+            if (speedSum <= 0.001f || _activeLegs.Count <= 0)
+            {
+                playerMover.BaseWalkSpeed = 0.8f;
+                playerMover.BaseSprintSpeed = 2.0f;
+            }
+            else
+            {
+                // Extra legs stack diminishingly.
+                // Final speed = speed sum/(leg count-log4(leg count))
+                playerMover.BaseWalkSpeed =
+                    speedSum / (_activeLegs.Count - (float) Math.Log(_activeLegs.Count, 4.0));
+
+                playerMover.BaseSprintSpeed = playerMover.BaseWalkSpeed * 1.75f;
+            }
+        }
+
+        /// <summary>
+        ///     Called when the layout of this body changes.
+        /// </summary>
+        private void OnBodyChanged()
+        {
+            // Calculate move speed based on this body.
+            if (Owner.HasComponent<MovementSpeedModifierComponent>())
+            {
+                _activeLegs.Clear();
+                var legParts = Parts.Values.Where(x => x.HasProperty(typeof(LegProperty)));
+
+                foreach (var part in legParts)
+                {
+                    var footDistance = DistanceToNearestFoot(part);
+
+                    if (Math.Abs(footDistance - float.MinValue) > 0.001f)
+                    {
+                        _activeLegs.Add(part, footDistance);
+                    }
+                }
+
+                CalculateSpeed();
+            }
+        }
+
+        /// <summary>
+        ///     Returns the combined length of the distance to the nearest <see cref="BodyPart"/> with a
+        ///     <see cref="FootProperty"/>. Returns <see cref="float.MinValue"/>
+        ///     if there is no foot found. If you consider a <see cref="BodyManagerComponent"/> a node map, then it will look for
+        ///     a foot node from the given node. It can
+        ///     only search through BodyParts with <see cref="ExtensionProperty"/>.
+        /// </summary>
+        public float DistanceToNearestFoot(IBodyPart source)
+        {
+            if (source.HasProperty<FootProperty>() && source.TryGetProperty<ExtensionProperty>(out var property))
+            {
+                return property.ReachDistance;
+            }
+
+            return LookForFootRecursion(source, new List<BodyPart>());
+        }
+
+        private float LookForFootRecursion(IBodyPart current,
+            ICollection<BodyPart> searchedParts)
+        {
+            if (!current.TryGetProperty<ExtensionProperty>(out var extProperty))
+            {
+                return float.MinValue;
+            }
+
+            // Get all connected parts if the current part has an extension property
+            if (!TryGetPartConnections(current, out var connections))
+            {
+                return float.MinValue;
+            }
+
+            // If a connected BodyPart is a foot, return this BodyPart's length.
+            foreach (var connection in connections)
+            {
+                if (!searchedParts.Contains(connection) && connection.HasProperty<FootProperty>())
+                {
+                    return extProperty.ReachDistance;
+                }
+            }
+
+            // Otherwise, get the recursion values of all connected BodyParts and
+            // store them in a list.
+            var distances = new List<float>();
+            foreach (var connection in connections)
+            {
+                if (!searchedParts.Contains(connection))
+                {
+                    continue;
+                }
+
+                var result = LookForFootRecursion(connection, searchedParts);
+
+                if (Math.Abs(result - float.MinValue) > 0.001f)
+                {
+                    distances.Add(result);
+                }
+            }
+
+            // If one or more of the searches found a foot, return the smallest one
+            // and add this ones length.
+            if (distances.Count > 0)
+            {
+                return distances.Min<float>() + extProperty.ReachDistance;
+            }
+
+            return float.MinValue;
+
+            // No extension property, no go.
+        }
     }
 }
